Add solution for problem `P - Area of a Star` Link: https://codeforces.com/contest/630/problem/P Tags: geometry, *2100 Ref: https://codeforces.com/contest/630/submission/206394935

+++ codeforces/630/P.cpp
@@ -55,9 +55,9 @@ double calculateTriangleArea(double angleA, double angleB, double sideLength) {
     return area;
 }
 void solve() {
-    double n, r;
+    ll n, r;
     cin >> n >> r;
-    double CA = 360/n;
+    double CA = 360.0/n;
     double firstA = CA/4;
     double secondA = CA/2;
     double area = calculateTriangleArea(firstA,secondA,r)*2;
